Rearrange order in which tokens are matched
Due to the greedy nature of attoparsec's string parser (specifically, it
does a "read n" and then a full string comparison), if we do not order
the matched tokens in order of increasing length, we would get partial
results.

This patch works around this behaviour by providing the <|> matches in
increasing (or equal) length.