Implements the choice of different signature types
Since I2P supports many different signature types, but not all nodes may
support all signature types, and some of them are not as good as the
others, we now provide the choice of signature type selection to the
client.

This also implements test cases which tests the support of all these
signature types.